+++ server_test.go
@@ -1,0 +1,7 @@
+package main
+
+import "testing"
+
+func TestStartServer(t *testing.T) {
+	StartServer()
+}
